pyproject.toml
[project]
name = "pdf-to-image"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymupdf>=1.24.0",
    "pillow>=11.3.0",
    "tkinterdnd2>=0.4.3",
    "customtkinter>=5.2.0",
]
